+++ tests/test_visitors/test_ast/test_annotations/test_argument_annotations.py
@@ -32,7 +32,7 @@ def function(
 ): ...
 """
 
-# Incorrect:
+# Wrong:
 
 wrong_multiline_arguments = """
 def function(